change type

+++ internal/core/dto/response/systemdto/system.go
@@ -1,7 +1,7 @@
 package systemdto
 
 type CalculationSystem struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float32 `json:"accrual"`
 }
